🚀 feat (Z06): Z06 series launched!
Z06 series is explained in updated readme.md file. This launch includes 3x2 size box and 1mm to 10mm drill bit lids for those.
I spent a lot of time on fine tuning this design, especially figuring out how to remove the lid when all adjacent spaces are also occupied.
In the end the best solution was to have angled feature in the lid parts, which allows the "top" of the lid to be pushed. This lifts the bottom
part of the lid and allows easy grabbing for removal.